requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.9.0
python-engineio==4.5.1
gevent==23.9.1
gevent-websocket==0.10.1 